requirements.txt
click==6.6
coreapi==2.1.1
coreapi-cli==1.0.6
Django==1.10.4
django-extensions==1.7.5
djangorestframework==3.5.3
httpie==0.9.6
itypes==1.1.0
Markdown==2.6.7
openapi-codec==1.2.1
Pygments==2.1.3
requests==2.12.3
six==1.10.0
uritemplate==3.0.0
wheel==0.24.0
